pyproject.toml
[tool.poetry]
name = "torch-ecg-volta"
version = "0.1.0"
description = ""
authors = ["lmaheo <lucas.maheo@gmail.com>"]
readme = "README.md"
packages = [{ include = "torch_ecg_volta" }]

[tool.poetry.scripts]
torch-ecg-server = "torch_ecg_volta.api.server:main"
hydra = "torch_ecg_volta.experiments.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.3"
scipy = "^1.11.4"
cython = "^3.0.7"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
wfdb = "3.2.0"
biosppy = "^2.1.0"
torch = "^2.1.2"
torchsummary = "^1.5.1"
tensorboardx = "^2.6.2.2"
tqdm = "^4.66.1"
deprecated = "^1.2.14"
deprecate-kwargs = "^0.0.4"
pyedflib = "^0.1.36"
pywavelets = "^1.5.0"
torch-optimizer = "^0.3.0"
requests = "^2.31.0"
xmltodict = "^0.13.0"
einops = "^0.7.0"
matplotlib = "^3.8.2"
bib-lookup = ">=0.0.19"
typing-extensions = "^4.9.0"
fastapi = "^0.108.0"
pydantic-settings = "^2.1.0"
uvicorn = "^0.25.0"
hydra-core = "^1.3.2"
mlflow = "^2.9.2"

[tool.poetry.group.dev.dependencies]
black = "22.8.0"
flake8 = "^7.0.0"
pytest = "^7.4.4"
pytest-xdist = "^3.5.0"
pytest-cov = "^4.1.0"
pcg-springer-features = "^0.3"
librosa = "^0.10.1"
torchaudio = "^2.1.2"
torch-audiomentations = "^0.11.0"
tensorboard = "^2.15.1"
openpyxl = "^3.1.2"
pre-commit = "^3.6.0"
sphinx = "<7.0"
nbsphinx = "^0.9.3"
sphinx-theme = "^1.0"
pydata-sphinx-theme = "^0.15.1"
sphinx-rtd-theme = "^2.0.0"
sphinx-book-theme = "^1.1.0"
sphinx-autobuild = "^2021.3.14"
sphinx-markdown-tables = "^0.0.17"
sphinx-copybutton = "^0.5.2"
sphinx-multiversion = "^0.2.4"
sphinx-design = "^0.5.0"
sphinx-emoji-favicon = "^0.2"
sphinxcontrib-napoleon = "^0.7"
sphinxcontrib-tikz = "^0.4.16"
sphinxcontrib-bibtex = "^2.6.1"
numpydoc = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autoflake]
check = false

[tool.isort]
profile = "black"
src_paths = ["src"]
line_length = 140

[tool.black]
line-length = 140
preview = true  # To enable string wrapping
